+++ kyumap/next.config.js
@@ -1,6 +1,13 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: false,
+  async rewrites() {
+    return [
+      {
+        source: "/upload/:slug",
+        destination: `${process.env.NEXT_PUBLIC_BASE_URL}/upload/:slug`,
+      },
+    ];
+  },
 };
 
 module.exports = nextConfig;
