--- kyumap/src/app/(afterLogin)/@modal/(.)detail/[postId]/_component/modalPost.tsx
@@ -1,8 +1,0 @@
-import styles from "../detail.module.css";
-
-type Props = {
-  isWide: boolean;
-  currentNumber: number;
-};
-
-export default function modalPost({ isWide, currentNumber }: Props) {}
