+++ kyumap/src/model/User.ts
@@ -1,5 +1,8 @@
+import { Post } from "./Post";
+
 export interface User {
   id: string;
   nickname: string;
   image: string;
+  posts: Post[];
 }
